+++ flutter/realm_flutter/ios/Classes/RealmPlugin.m
@@ -43,6 +43,7 @@ void dummy(void) {
   realm_list_size(NULL, 0);
   realm_results_snapshot(NULL);
   realm_http_transport_new(NULL, NULL, NULL);
+  realm_config_set_should_compact_on_launch_function(NULL, NULL, NULL);
   realm_app_credentials_new_anonymous();
   realm_app_config_new(NULL, NULL);
   realm_sync_client_config_new();